search-button now links to a new page with the url of the pokemon searched. The api call is triggered on the new page. I've moved the css for that section into its own file into that page.
