+++ src/App.css
@@ -1,99 +1,53 @@
 .App {
-    display: flex;
-    justify-content: center;
-    flex-direction: column;
-    font-family: Arial, Helvetica, sans-serif;
-    font-size: 1.8rem;
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+  font-family: Arial, Helvetica, sans-serif;
+  font-size: 1.8rem;
 }
 
 .title-section {
-    display: flex;
-    background-color: #5db9ff;
-    width: 100%;
-    height: 300px;
-    color: white;
-    flex-direction: column;
-    align-items: center;
+  display: flex;
+  background-color: #5db9ff;
+  width: 100%;
+  height: 270px;
+  color: white;
+  flex-direction: column;
+  align-items: center;
+}
+
+.title {
+  font-size: 3rem;
+  color: white;
 }
 
 .search-input {
-    width: 200px;
-    height: 40px;
-    border-radius: 10px;
-    border: 1px solid #51a0dd;
-    font-size: 1rem;
-    padding-left: 8px;
+  width: 200px;
+  height: 40px;
+  border-radius: 10px;
+  border: 1px solid #51a0dd;
+  font-size: 1rem;
+  padding-left: 8px;
 }
 
 .search-button {
-    width: 180px;
-    height: 60px;
-    margin: 1em;
-    border-radius: 10px;
-    font-size: 1rem;
-    padding-left: 8px;
-    border: none;
-    background-color: #407faf;
-    color: white;
+  width: 180px;
+  height: 60px;
+  margin: 1em;
+  border-radius: 10px;
+  font-size: 1rem;
+  padding-left: 8px;
+  border: none;
+  background-color: #407faf;
+  color: white;
 }
 
 .search-button:hover {
-    cursor: pointer;
+  cursor: pointer;
 }
 
-.display-section {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.pokemon-image {
-    margin-top: -1em;
-    height: 300px;
-}
-
-.stats-window {
-    background-color: #fbd743;
-    border-radius: 2em;
-    padding-top: 1em;
-    padding-right: 0em;
-    padding-bottom: 0em;
-    padding-left: 1em;
-    margin-top: 1em;
-}
-
-.stats-title {
-    text-align: center;
-    margin-top: 0;
-    margin-right: 0.4em;
-    margin-bottom: 0.4em;
-    margin-left: 0.4em;
-    font-size: 2.4rem;
-    color: darkslategrey;
-}
-
-.text-and-image-wrapper {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-}
-
-.text-stats {
-    align-items: center;
-    background-color: #ebc324;
-    border-radius: 10px;
-    height: fit-content;
-    padding: 8px;
-    margin-top: 2em;
-}
-
-.text-stats p {
-    font-size: 0.8rem;
-    margin: 0;
-    padding-bottom: 4px;
-    color: rgb(62, 62, 62);
-}
-
-.bold {
-    font-weight: bold;
+.call-to-action {
+  text-align: center;
+  font-size: 3rem;
+  color: black;
 }
